[LCR-238] refactor: OpenFeign 사용하여 오늘 예약자 수 조회 기능 리팩토링

+++ src/main/java/com/lgcns/domain/reservationStats/dto/response/ReservationStatsResponse.java
@@ -4,12 +4,12 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.List;
 
 public record ReservationStatsResponse(
-        @Schema(description = "금일 예약 수", example = "100") Integer reservedCount,
+        @Schema(description = "금일 예약 수", example = "100") Long reservedCount,
         @Schema(description = "요일별 예약 수", example = "[{'day': '월', 'reservedCount': 10},...]")
                 List<DayOfWeekReservationCountResponse> chart) {
 
     public static ReservationStatsResponse of(
-            Integer reservedCount, List<DayOfWeekReservationCountResponse> chart) {
+            Long reservedCount, List<DayOfWeekReservationCountResponse> chart) {
         return new ReservationStatsResponse(reservedCount, chart);
     }
 }
